refactor: :fire: updated tablet design and breakpoint usage
