+++ src/components/custom-tabs/index.ts
@@ -1,0 +1,3 @@
+import CustomTabs from './custom-tabs';
+
+export default CustomTabs;
