+++ src/features/dashboard/components/api-token-table/api-table.module.scss
@@ -1,3 +1,5 @@
+@use 'src/styles/mixins' as *;
+
 .flex-center {
   display: flex;
   flex-direction: column;
@@ -8,7 +10,7 @@
   align-self: self-end;
   position: relative;
 
-  @media (max-width: 1023px) {
+  @include tablet-lg {
     align-self: center;
   }
 
@@ -23,7 +25,7 @@
   border: 1px solid var(--opacity-black-100);
   border-radius: 32px;
 
-  @media screen and (max-width: 1023px) {
+  @include tablet-lg {
     border: none;
     margin: 0 1rem;
     gap: 3rem;
@@ -93,7 +95,7 @@
     padding: 48px 48px 25px;
     gap: 16px;
 
-    @media (max-width: 1023px) {
+    @include tablet-lg {
       flex-direction: column;
       align-items: center;
       text-align: center;
@@ -105,7 +107,7 @@
     }
 
     &__button {
-      @media screen and (max-width: 1023px) {
+      @include tablet-lg {
         margin-top: 1rem;
       }
 
@@ -117,7 +119,7 @@
       &__text {
         font-size: 16px;
 
-        @media screen and (max-width: 786px) {
+        @include mobile {
           padding: 0 1rem;
         }
       }
@@ -129,7 +131,7 @@
       justify-content: space-between;
       align-self: stretch;
 
-      @media (max-width: 1023px) {
+      @include tablet-lg {
         flex-direction: column;
         align-items: center;
         text-align: center;
