+++ src/features/dashboard/components/common-table/cell-scopes.module.scss
@@ -1,16 +1,20 @@
 @use 'src/styles/utility' as *;
+@use 'src/styles/mixins' as *;
 
 .scope {
   display: inline-block;
-  border: rem(0.1) solid var(--ifm-color-emphasis-400);
+  border: 1px solid var(--ifm-color-emphasis-400);
   border-radius: 4px;
-  padding: rem(0.2) rem(0.8);
-  font-size: rem(1.2);
-  margin: rem(0.5);
+  padding-inline: 4px;
+  font-size: 12px;
+  margin: 4px;
   color: var(--ifm-color-emphasis-700);
+  @include mobile {
+    padding-block: 2px;
+  }
 }
 
 .adminScope {
-  border: rem(0.1) solid var(--admin-border);
+  border: 1px solid var(--admin-border);
   color: var(--admin-text);
 }
