+++ src/hooks/useDeviceType/index.tsx
@@ -14,7 +14,7 @@ const useDeviceType = (): TUseDeviceType => {
     const handleResize = () => {
       if (window.matchMedia('(max-width: 768px)').matches) {
         setDeviceType('mobile');
-      } else if (window.matchMedia('(max-width: 1023px)').matches) {
+      } else if (window.matchMedia('(max-width: 1024px)').matches) {
         setDeviceType('tablet');
       } else {
         setDeviceType('desktop');
